+++ tests/nativeapp/utils.py
@@ -89,20 +89,24 @@ def mock_execute_helper(mock_input: list):
     return side_effects, expected
 
 
+# TODO: move to shared utils between integration tests and unit tests once available
 def touch(path: str):
     file = Path(path)
     file.parent.mkdir(exist_ok=True, parents=True)
     file.write_text("")
 
 
-def stringify(p: Path):
+# Helper method, currently only used within assert_dir_snapshot
+def _stringify_path(p: Path):
     if p.is_dir():
         return f"d {p}"
     else:
         return f"f {p}"
 
 
-def all_paths_under_dir(root: Path) -> List[Path]:
+# Helper method, currently only used within assert_dir_snapshot.
+# For all other directory walks in source code, please use available source utils.
+def _all_paths_under_dir(root: Path) -> List[Path]:
     check = os.getcwd()
     assert root.is_dir()
 
@@ -118,11 +122,12 @@ def all_paths_under_dir(root: Path) -> List[Path]:
     return sorted(paths)
 
 
+# TODO: move to shared utils between integration tests and unit tests once available
 def assert_dir_snapshot(root: Path, snapshot) -> None:
-    all_paths = all_paths_under_dir(root)
+    all_paths = _all_paths_under_dir(root)
 
     # Verify the contents of the directory matches expectations
-    assert "\n".join([stringify(p) for p in all_paths]) == snapshot
+    assert "\n".join([_stringify_path(p) for p in all_paths]) == snapshot
 
     # Verify that each file under the directory matches expectations
     for path in all_paths:
